Updated profile page
I added a link in the header bar dropdown menu for the user's profile page.

For a user with role id < 7:
I made the icon and role id uneditable since users below role id below 7 should not be able to edit those.

For a user with role id >= 7:
I added a dropdown menu with all of the names of the font awesome icons with a spot that previews the selected icon.