Added membership expiration to user profile
Added a section to the user profile that shows when the user's membership ends. If there is no membership expiration specified in the database, it will display indefinite instead.

Extra: I restructured the div tag structure in the profile page. Also, I add a user profile link to the sidebar for users with role id >= 7.