Modified profile page
I made the role ID uneditable since there would not be a reason for the user to change their role ID. Because of that, I removed the code for dealing with a POST that contains role ID info.

I also made the notes section of the profile page only display if the user has a role ID >= 7.